pkgconfig: Move ligkmod features from kmod.pc to libkmod.pc
sysconfdir, distconfdir, module_compressions, and module_signatures are
libkmod features, these should be in libkmod.pc, not kmod.pc.

module_directory is special: While libkmod has a default for when the
directory is not specified depmod does spefcify it, and has its own
embetted value.

Technically it's possible to have kmod and libkmod with mismatched
values due to them being built separately. Also on a binary OS
distribution the libkmod.pc would be installed only with libkmod
development headers, not the tool.

Signed-off-by: Michal Suchanek <msuchanek@suse.de>