CONTRIBUTING.md: mention s-o-b tag, includ DCO copy
While CODING-STYLE states that using Signed-off-by is an error, a large
number of contributors have been using it. Most, if not all, of those
are people who have contributed to the kernel, and by extension know
about the DCO.

So let's add an in-tree copy of the file, explicitly document what is
meant with the tag (other projects use if for different purposes) and
drop the note from CODING-STYLE.

I have not went full bananas to require it, since based on the numbers
below, majority of commits lack it. We could reconsider that at later
point, if needed.

Numbers, as of commit 11ccabd ("libkmod: document the symbols file")
 - total number of commits - 1640
 - number of commits with at least one s-o-b - 357
   - this can be s-o-b by the author, maintainer and/or both
 - total number of contributors (bots including) - 109
 - number of individuals with at least one s-o-b - 42

v2
 - drop html bits from DCO
 - tag -> trailer

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>
Link: https://github.com/kmod-project/kmod/pull/134
Signed-off-by: Lucas De Marchi <lucas.de.marchi@gmail.com>