libkmod: Introduce fallback module directory
Currently, traditional Linux systems using initramfs are forced to keep two
copies of the modules which may be needed both during early boot and for
normal system operation. This seems to be inevitable for the case when
initramfs is built locally but with the rise of distro shipped UKIs the
question whether it would be possible to avoid the redundancy arises.

To make it possible to reuse modules from initramfs without copying them to
the root filesystem (which may be read-only), the suggested approach is to
preserve them on tmpfs, e.g. in /run/modules/`uname -r` but the challenge
is to make kmod search there. It would certainly be possible to e.g. bind
mount /run/modules/`uname -r` to /lib/modules/`uname -r` but this will make
the destination read-only and traditional tools which are used to install
packages with modules (e.g. rpm) will break. Instead, a 'fallback'
directory approach is suggested: in case main /lib/modules/`uname -r` is
completely missing, kmod tools will check whether /run/modules/`uname -r`
exist and search there.

In theory, it would've been possible to replace this fallback option with a
dedicated config for kmod making modules location dynamic but this seems to
be a more intrusive change and it will also require additional external
actions to switch from 'fallback' to 'main' if/when modules get installed
to the main system.

Note, no merging of /run/modules/`uname -r` and /lib/modules/`uname -r` is
performed. It is expected, that in case /lib/modules/`uname -r` exists, it
is a superset and modules from initramfs are no longer needed.

Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>