tools/depmod: cast variable in DBG() statement
Older versions of clang (14.0.0 as seen in Ubuntu 22.04) will flag a
warning that we're using unsigned int, with a unsigned char modifier.

Newer versions like 16.0.6 (Debian unstable), 17.0.6 (Alpine), etc do
not flag this as an issue.

In practise the number fits in the latter range, so just add a cast to
silence the warning.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>
Link: https://github.com/kmod-project/kmod/pull/172
Signed-off-by: Lucas De Marchi <lucas.de.marchi@gmail.com>