XXX: Enable ELFDBG with debug-messages
In my early days of hacking at kmod, I was really confused why none of
the ELFDBG() logging worked... Recently we discovered the code path was
broken and didn't build.

Currently some distributions --enable-debug for kmod (Debian and
Fedora), with which others (Gentoo and Yocto) leaving the final decision
to the builder.

As of earlier commit, distributions no longer need --enable-debug to get
the log priority/severity logged so in theory some of them will drop the
toggle. Even if they keep it, the extra messages are available as one
explicitly raises the log level.

Since the ELFDBG() messages can spam the (sys)log, add a configure-time
warning for builders to see and make their informed decision.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>