Mandate logging, remove built-time toggle
Currently we have the libkmod logging behind a build-time switch, which
defaults to ON. Where all the distributions keep the default AFAICT.

If one is to switch it off, then the libkmod log API becomes
non-functional. In addition, the tools' syslog facility also does
nothing...

The icky part is that the user (API or tools) is not informed about
this and the "kmod --version" (which otherwise lists features), does not
mention logging at all.

Just remove the toggle, so that people cannot shoot themselves in the
foot and we cut down on the build permutations.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>