libkmod: used unsigned type for INDEX_CHILDMAX and co
Currently the define is an signed int of 128, which we store in a char
triggering a clang warning. This is expected since the sign of char is
platform specific.

Convert the macro to unsigned and the respective variables to unsigned
char, alongside the documentation comment.

With that done, adjust the first/last checks so we don't get
tautological-compare warnings from clang.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>