ci: disable sanitizers for ubuntu 22.04
It seems that the meson shipped with Ubuntu 22.04 is too old (buggy?) to
set the ASAN_OPTIONS and similar variables. Which means that we don't
have an easy way to detect if the sanitizers are enabled in the upcoming
test wrapper script.

The alternative is to grab newer meson, either via pip or otherwise.
Although considering we have decent coverage via the other distributions
it doesn't seem worth the hassle.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>