libkmod: Allow better optimization
If we have native endianess, i.e. parsing modules for the running
system, assist the compiler to note that it is really much faster to
move a word/qword etc. instead of actually running through a loop.

Reduces library instructions on x86_64 by 1.4 % and binary instructions
by 3 % with default configuration.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>