docs: annotate the deprecated API
The kmod_module_get_filtered_blacklist() was deprecated since kmod v6 in
favour of kmod_module_apply_filter().

Add the decoration so gtk-doc includes it in the generated html and also
add a designated index.

It seem that gtk-doc insists on having deprecation guards, so not it
prints an extra warning like:

  warning: XXX is deprecated in the inline comments, but no deprecation
  guards were found around the declaration. (See the --deprecated-guards
  option for gtkdoc-scan.)

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>