libkmod: store common section off/size and use them
Currently, we repeatedly loop over the elf headers looking for five well
known sections. Just do it once in kmod_elf_new() and reuse the data as
needed.

Note, that not all sections are guaranteed to be available, so calloc()
the struct kmod_elf, so we don't rely on garbage data later on.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>