README: Add release/distro/end-user recommendation
This may be useful for anyone trying a non-debug build and serve as
baseline for distros. Even for developers, when benchmarking kmod, it's
better to run a release-oriented kmod rather than the slow one due to
all debug/sanitizers.

By keeping the -D in the command line rather than in a ini file,
we also guarantee meson shows it in the summary, regardless of
https://github.com/mesonbuild/meson/issues/13865.

Closes: https://github.com/kmod-project/kmod/issues/220
Signed-off-by: Lucas De Marchi <lucas.de.marchi@gmail.com>