ci/codeql: use Ubuntu 24.04, adjust build flags
Recently the CI has been having issues with Ubuntu 22.04, where the
installations of tzdata was blocking in interactive mode. Bump it to
24.04 which doesn't seem to have this issue.

While in here drop the docs and manpages flags - both of those combined
take ~1s of CI time - and disable the sanitizers and tests instead.

This means we no longer get pre-existing code problems in the tests,
which is fine considering the number of pre-existing issues in there.
Another nice side effect is that the pipeline time is reduced by ~50%.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>