XXX: testsuite: make all output.files relative to TC_ROOTFS
With this commit all our tests have a single entrypoint (from a
file-system POV) which means we get a lot less duplication, all test
specific files (be that config, input, expectations, etc) are in
designated folders.

As result, this paves the route to running the tests in chroot-like
enviroment were we can drop-in the folder, adjust the module rootfs
location in path.so or equivalent and have the test(s) churn.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>