ci: Use include to completly specify the matrix
The permutation available in github and extending it "include" or
reducing it with "exclude" are much harder than needed. It doesn't seem
we can have a mix of "properties" with the configurations. Just give up
and completly specify the matrix. Now that there's only 1 build system,
at least doing all the permutations doesn't make it too big.

Signed-off-by: Lucas De Marchi <lucas.de.marchi@gmail.com>