tools: use space for long opts
The getopt_long() API handles both space and = separated long options
just fine. On the other hand, supporting both in the shell completions
means that we'll have to quadruple them in terms of size/complexity.

So let's try to keep things simple and consistent, by recommending only
the space variant.

Make sure to update both the tools' --help and the respective man pages.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>