Introduce KMOD_STRERROR to ensure usage is safer.
While recent strerror implementations are thread aware/safe,
old ones are not, and the issue of shared library safety remains.

That's why strerrordesc_np() was added that is thread safe,
async-signal-safe (does not call malloc) and there is no data
race with other functions.

Use that if available, else fallback to the closest that is
calling strerror_r with a buffer whose scope expires inmediately.

This concerns only libkmod, in the future the toolchain may
implement per-dso allocations to make the problem disappear in practice.

Note that strerrordesc_np does not display error in the current locale
but only in english, this is an small price to pay for safety.