Remove dietlibc stat::st_mtim workaround
Some versions of dietlibc lack struct stat::st_mtim, which was
introduced with POSIX.1-2008.

Considering the slow development cycle of dietlibc (last three releases
are 2024, 2018 and 2013) and there's no popular user (nor one in CI) of
it, let's remove the workaround.

All the other runtimes that we target - glibc, bionic and musl - have
the struct member.

Iff needed we can reintroduce the workaround at a later stage alongside
a CI permutation.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>