meson: disable Wdeclaration-after-statement
Having declarations after statement is a C90 requirement, which was
relaxed with C99. While there are some arguments for having it enabled
in new code, there are benefits to the contrary which are more appealing
IMHO.

Namely, by keeping the declarations where the variables are used, better
outlines the scope, reduces chances of shadowing and helps us avoid
"unused variable" compiler warning in some paths.

Flip the warning to -Wno for now and future patches can remove some of
the workarounds we have in-tree.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>