tools/modprobe: standardize on --show-foo, soft-deprecate the rest
Currently we have a range of options which show specific details,
although they vary in prefix or spelling:
 - prefix: dump vs resolve vs show
 - with and without dash after show

Converge on --show-foo and soft deprecate the rest. As in, the old
options are still accepted but only the (new) consistent names are
listed in --help and the manual page.

To avoid any confusion with --show (aka --dry-run) the option is also
soft-deprecated.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>