clang-format: update attribute macros
Use the command to update the attribute macros list. Seemingly we've
updated the list manually, which meant that a) some symbols were not
annotated the expected way and b) some macros were not picked up.

First we start with auto-generating the list, the other issues will be
handled as follow-ups.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>