testsuite: fixup argument handling
Currently we can pass any number of tests as arguments - be that in
oneshot mode or not.

At the same time, oneshot requires only one test while normal mode can
have up-to one.

Fix that up and as a bonus we no longer need the exit() in
test_run_spawned() \o/

While in here, add a few comments around our arg/argc handling.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>