libkmod: remove unreachable assert() instances
In a handful of places, we're using asserts to guard for internal API
misuse. At the same time, having asserts() in a library is generally a
bad idea, especially since distributions may be building with or without
"-DNDEBUG".

Based on code analysis, some instances are not reachable so let's remove
them.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>