libkmod: remove assert() from public API
Handful of public facing functions check if the respective pointers are
NULL initialized, while others assert() on it. Remove the latter (which
is frowned upon in shared libraries) and return early in that case.

Should this change yield any user-facing changes, this means they were
already getting random corruption and haven't noticed it yet.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>