shared/strbuf: remove remaining assert() instances
Remove the final assert calls from the library and the respective test.
The latter being the only xfail test in the whole project and covering only
one of the asserts being removed.

In addition, since assert() uses abort() the code coverage of the
function is completely thrown away... Which isn't great.

We could reintroduce some of these checks (asserts remove here and
previous commits) and/or tests, although it should probably be done
differently. Having any assert() in a public/shared library is a no-go.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>